+++ CTFd/utils/crypto/__init__.py
@@ -1,4 +1,6 @@
+from CTFd.utils import string_types
 from passlib.hash import bcrypt_sha256
+import hashlib
 
 
 def hash_password(plaintext):
@@ -7,3 +9,9 @@ def hash_password(plaintext):
 
 def verify_password(plaintext, ciphertext):
     return bcrypt_sha256.verify(plaintext, ciphertext)
+
+
+def sha256(p):
+    if isinstance(p, string_types):
+        p = p.encode('utf-8')
+    return hashlib.sha256(p).hexdigest()
