+++ CTFd/utils/updates/__init__.py
@@ -3,7 +3,7 @@ from flask import current_app as app
 from CTFd.utils import get_config, set_config, get_app_config
 from CTFd.utils.config import is_setup
 from CTFd.models import db, Challenges, Users, Teams
-from CTFd.utils.security.passwords import sha256
+from CTFd.utils.crypto import sha256
 from platform import python_version
 import requests
 import time
